feat : ViewController에서 Clustering으로 옮기고 영역표시 추가
